json-ast/kotlin: expose AST helpers
